+++ src/main/java/com/github/manolo8/darkbot/utils/LogUtils.java
@@ -1,12 +1,8 @@
 package com.github.manolo8.darkbot.utils;
 
-import org.jetbrains.annotations.NotNull;
-
 import java.io.BufferedOutputStream;
-import java.io.Closeable;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
-import java.io.FilterOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.PrintStream;
@@ -140,129 +136,4 @@ public class LogUtils {
         }
     }
 
-    public static class MultiOutputStream extends FakeMultiOutputStream {
-
-        OutputStream second;
-
-        public MultiOutputStream(OutputStream out, OutputStream second) {
-            super(out);
-            this.second = second;
-        }
-
-        @Override
-        public synchronized void write(int b) throws IOException {
-            super.write(b);
-            second.write(b);
-        }
-
-        @Override
-        public synchronized void write(byte @NotNull [] b) throws IOException {
-            super.write(b);
-            second.write(b);
-        }
-
-        @Override
-        public synchronized void write(byte @NotNull [] b, int off, int len) throws IOException {
-            super.write(b, off, len);
-            second.write(b, off, len);
-        }
-
-        @Override
-        public synchronized void flush() throws IOException {
-            super.flush();
-            second.flush();
-        }
-
-        @Override
-        public synchronized void close() throws IOException {
-            try {
-                super.close();
-            } finally {
-                second.close();
-            }
-        }
-    }
-
-    public static class FakeMultiOutputStream extends FilterOutputStream {
-
-        public FakeMultiOutputStream(final OutputStream proxy) {
-            super(proxy);
-        }
-
-        @SuppressWarnings("unused") // Possibly thrown from subclasses.
-        protected void afterWrite(final int n) throws IOException {
-            // noop
-        }
-
-        @SuppressWarnings("unused") // Possibly thrown from subclasses.
-        protected void beforeWrite(final int n) throws IOException {
-            // noop
-        }
-
-        @Override
-        public void close() throws IOException {
-            close(out);
-        }
-
-        @Override
-        public void flush() throws IOException {
-            try {
-                out.flush();
-            } catch (final IOException e) {
-                handleIOException(e);
-            }
-        }
-
-        protected void handleIOException(final IOException e) throws IOException {
-            throw e;
-        }
-
-        @Override
-        public void write(final byte @NotNull [] bts) throws IOException {
-            try {
-                final int len = length(bts);
-                beforeWrite(len);
-                out.write(bts);
-                afterWrite(len);
-            } catch (final IOException e) {
-                handleIOException(e);
-            }
-        }
-
-        @Override
-        public void write(final byte @NotNull [] bts, final int st, final int end) throws IOException {
-            try {
-                beforeWrite(end);
-                out.write(bts, st, end);
-                afterWrite(end);
-            } catch (final IOException e) {
-                handleIOException(e);
-            }
-        }
-
-        @Override
-        public void write(final int idx) throws IOException {
-            try {
-                beforeWrite(1);
-                out.write(idx);
-                afterWrite(1);
-            } catch (final IOException e) {
-                handleIOException(e);
-            }
-        }
-
-        public static void close(final Closeable closeable) throws IOException {
-            if (closeable != null) {
-                try {
-                    closeable.close();
-                } catch (final IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-
-        public static int length(final byte[] array) {
-            return array == null ? 0 : array.length;
-        }
-    }
 }
